ditto

+++ src/main/java/com/higgs/staged/StagedActor.java
@@ -55,7 +55,7 @@ public abstract class StagedActor implements MouseReactant, KeyReactant {
         return Optional.ofNullable(this.animation).map(Animation::getFrame).map(BufferedImage::getHeight).orElse(-1);
     }
 
-    public void setX(final int x) {
+    public void setX(final double x) {
         this.x = x;
     }
 
@@ -63,7 +63,7 @@ public abstract class StagedActor implements MouseReactant, KeyReactant {
         return this.x;
     }
 
-    public void setY(final int y) {
+    public void setY(final double y) {
         this.y = y;
     }
